// CuerpoSanoMantenimiento/src/app/models/CSE_Ejercicio.ts
export class CSE_Ejercicio{

    Id?: number;
    NombreEjercicio: string;
    Descripcion: string;
    Recomendaciones: string;
    MusculoTrabajado: string;
    MusculosSecundario: string;
    UrlFoto: string;
    UrlVideo: string;
    RepeticionesRecomendadas: string;
    TiempoRecomendado: number;
}